// src/types/saving.tsx
import {TextStyle, ViewStyle, ImageStyle} from 'react-native';

export interface ISavingsStyles {
	mainView: ViewStyle;
	marginRight10: ViewStyle;
	marginBottom10: ViewStyle;
	totalAvailableCash: TextStyle;
	moneyBig: TextStyle;
	moneyLittle: TextStyle;
	littleGreyText: TextStyle;
	listHeader: ViewStyle;
	inputButton: ViewStyle;
	buttonFilter: ViewStyle;
	styleImage: ImageStyle;
	imageContainer: ViewStyle;
	headerBackground: ViewStyle;
	viewInterestGoodness: ViewStyle;
	viewInterestGoodnessView: ViewStyle;
	viewInterestGoodnessText: TextStyle;
	viewInterestGoodnessAmount: TextStyle;
}

export interface ISavingsListStyles {
	listMoney: ViewStyle;
	viewList: ViewStyle;
	viewRowCenter: ViewStyle;
	colorGrey: TextStyle;
	flexRow: ViewStyle;
	width190: ViewStyle;
	marginRight5: ViewStyle;
	colorGreen: TextStyle;
	fontSize15: TextStyle;
	fontSize10: TextStyle;
	fontSize25: TextStyle;
	fontSize20: TextStyle;
	paddingVertical20: ViewStyle;
}